+++ includes/Modules/DirectCheckout/direct-checkout.asset.php
@@ -1,0 +1,5 @@
+<?php
+return [
+    'dependencies' => ['jquery'],
+    'version'      => '1.0.0',
+];